servers/whois-mcp/pyproject.toml
[project]
name = "whois-mcp"
version = "0.1.0"
description = "WHOIS and DNS lookup MCP server for the Cyber Operate platform"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.11.1",
    "mcp>=1.23.0",
    "uvicorn>=0.34.0",
    "starlette>=0.41.0",
    "httpx>=0.28.1,<1.0",
    "python-whois>=0.9.4",
    "dnspython>=2.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server.py"]
